Protected

+++ src/main/java/servlet/SaveLocation.java
@@ -1,27 +1,20 @@
 package servlet;
 
 import entitiy.Location;
-import factory.PersistorFactory;
-import manager.MockEntityManager;
-import manager.SQLEntityManager;
-
+import manager.PersistorFactory;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-/**
- * Created by MalindaK on 5/7/2018.
- */
 public class SaveLocation extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-
         try {
             String name = request.getParameter("name");
-            Float lat = Float.parseFloat(request.getParameter("lat"));
-            Float lon = Float.parseFloat(request.getParameter("lon"));
+            Double lat = Double.parseDouble(request.getParameter("lat"));
+            Double lon = Double.parseDouble(request.getParameter("lon"));
 
             Location location = new Location(name, lat, lon);
             PersistorFactory.getPersistor().add(location);
